backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
